+++ internal/ui/components/keyboard.go
@@ -107,6 +107,7 @@ func (a *App) setupKeyboardHandlers() {
 			a.pages.HasPage("resizeStorage") ||
 			a.pages.HasPage("profileWizard") ||
 			a.pages.HasPage("profileName") ||
+			a.pages.HasPage("pluginsManager") ||
 			a.pages.HasPage("contextMenu") ||
 			a.pages.HasPage("about") ||
 			a.pages.HasPage("snapshots") ||